+++ settings.gradle.kts
@@ -28,7 +28,6 @@ rootProject.name = "xKylee Plugins"
 include(":alchemicalhydra")
 include(":aoewarnings")
 include(":betterequipmentinspector")
-include(":blackjack")
 include(":castlewarsextended")
 include(":cerberus")
 include(":clanmanmode")
@@ -53,6 +52,7 @@ include(":hotkeytowalk")
 include(":hydra")
 include(":inferno")
 include(":kalphitequeen")
+include(":leftclickblackjack")
 include(":lizardmenshaman")
 include(":lootassist")
 include(":menuentryswapperextended")
